quick commit -_-

+++ config-example.h
@@ -8,5 +8,5 @@
 #define PORT 80
 
 static constexpr const char * ROOT_FILE = "compiled.html.gz"; 
-static constexpr uint32_t SD_SECTOR_SIZE = 512;
+static constexpr uint64_t SD_SECTOR_SIZE = 512;
